tests/integrated/test_bailian_fc_deploy/examples/custom_defined_app/requirements.txt
fastapi
uvicorn
python-dotenv
openai


